Not allow user to delete the using account

+++ company-registration/src/pages/StaffList/StaffTableRow.js
@@ -13,15 +13,20 @@ export default class CompanyTableRow extends Component {
 
     deleteStaff = (staff) => {
         console.log(staff)
-        axios.delete('http://localhost:4000/staff/' + staff.username,
-         {headers:{"x-access-token":this.props.userData.token}})
-        .catch((error) => {
-            console.log(error)
-        })
-        .then(res => {
-            console.log(res)
-            this.props.refresh()
-        })
+        if ( staff.username == this.props.userData.username) {
+            console.log("You cannot delete your own account!")
+        } else {
+            axios.delete('http://localhost:4000/staff/' + staff.username,
+            {headers:{"x-access-token":this.props.userData.token}})
+           .catch((error) => {
+               console.log(error)
+           })
+           .then(res => {
+               console.log(res)
+               this.props.refresh()
+           })
+        }
+        
     }
 
     render() {
@@ -37,9 +42,15 @@ export default class CompanyTableRow extends Component {
                     <Button size="sm" variant="outline-primary">View</Button>
                     </Link>
                     &nbsp;&nbsp; */}
+                    {this.props.obj.username == this.props.userData.username ?(
+                        <Button size="sm" onClick={() => {
+                            console.log('disabled')
+                        }}variant="outline-dark" disabled> Delete</Button>
+                    ): (
                     <Button size="sm" onClick={() => {
                         this.deleteStaff(this.props.obj)
-                    }}variant="outline-danger"> Delete</Button>
+                    }}variant="outline-danger"> Delete</Button>) }
+                    
                 </td>       
 
             </tr>
